+++ sdk/compute/arm-compute/src/models/parameters.ts
@@ -110,7 +110,7 @@ export const $host: OperationURLParameter = {
 export const apiVersion: OperationQueryParameter = {
   parameterPath: "apiVersion",
   mapper: {
-    defaultValue: "2021-04-01",
+    defaultValue: "2021-07-01",
     isConstant: true,
     serializedName: "api-version",
     type: {
@@ -479,6 +479,16 @@ export const forceDeletion: OperationQueryParameter = {
   }
 };
 
+export const hibernate: OperationQueryParameter = {
+  parameterPath: ["options", "hibernate"],
+  mapper: {
+    serializedName: "hibernate",
+    type: {
+      name: "Boolean"
+    }
+  }
+};
+
 export const statusOnly: OperationQueryParameter = {
   parameterPath: ["options", "statusOnly"],
   mapper: {
@@ -492,6 +502,7 @@ export const statusOnly: OperationQueryParameter = {
 export const skipShutdown: OperationQueryParameter = {
   parameterPath: ["options", "skipShutdown"],
   mapper: {
+    defaultValue: false,
     serializedName: "skipShutdown",
     type: {
       name: "Boolean"
@@ -858,12 +869,10 @@ export const runCommand1: OperationParameter = {
   mapper: VirtualMachineRunCommandUpdateMapper
 };
 
-export const apiVersion1: OperationQueryParameter = {
-  parameterPath: "apiVersion",
+export const includeExtendedLocations: OperationQueryParameter = {
+  parameterPath: ["options", "includeExtendedLocations"],
   mapper: {
-    defaultValue: "2019-04-01",
-    isConstant: true,
-    serializedName: "api-version",
+    serializedName: "includeExtendedLocations",
     type: {
       name: "String"
     }
@@ -886,7 +895,7 @@ export const diskName: OperationURLParameter = {
   }
 };
 
-export const apiVersion2: OperationQueryParameter = {
+export const apiVersion1: OperationQueryParameter = {
   parameterPath: "apiVersion",
   mapper: {
     defaultValue: "2020-12-01",
@@ -1019,18 +1028,6 @@ export const galleryName: OperationURLParameter = {
   mapper: {
     serializedName: "galleryName",
     required: true,
-    type: {
-      name: "String"
-    }
-  }
-};
-
-export const apiVersion3: OperationQueryParameter = {
-  parameterPath: "apiVersion",
-  mapper: {
-    defaultValue: "2020-09-30",
-    isConstant: true,
-    serializedName: "api-version",
     type: {
       name: "String"
     }
@@ -1194,7 +1191,7 @@ export const cloudServiceName: OperationURLParameter = {
   }
 };
 
-export const apiVersion4: OperationQueryParameter = {
+export const apiVersion2: OperationQueryParameter = {
   parameterPath: "apiVersion",
   mapper: {
     defaultValue: "2021-03-01",