+++ sdk/metricsadvisor/ai-metrics-advisor/test/public/datafeed.spec.ts
@@ -23,7 +23,7 @@ import {
   AzureLogAnalyticsDataFeedSource
 } from "../../src";
 import { createRecordedAdminClient, testEnv, makeCredential } from "./util/recordedClients";
-import { Recorder } from "@azure/test-utils-recorder";
+import { Recorder } from "@azure-tools/test-recorder";
 import { matrix } from "./util/matrix";
 
 matrix([[true, false]] as const, async (useAad) => {