+++ sdk/monitor/monitor-query/src/metricsQueryClient.ts
@@ -2,14 +2,14 @@
 // Licensed under the MIT license.
 import { TokenCredential } from "@azure/core-auth";
 import { PipelineOptions, bearerTokenAuthenticationPolicy } from "@azure/core-rest-pipeline";
+import { PagedAsyncIterableIterator } from "@azure/core-paging";
 
 import {
-  GetMetricDefinitionsOptions,
-  GetMetricDefinitionsResult,
-  GetMetricNamespacesOptions,
-  GetMetricNamespacesResult,
-  QueryMetricsOptions,
-  QueryMetricsResult
+  ListMetricDefinitionsOptions,
+  ListMetricNamespacesOptions,
+  MetricsQueryOptions,
+  MetricsQueryResult,
+  MetricDefinition
 } from "./models/publicMetricsModels";
 
 import {
@@ -22,6 +22,7 @@ import {
 } from "./generated/metricsdefinitions/src";
 import {
   KnownApiVersion20171201Preview as MetricNamespacesApiVersion,
+  MetricNamespace,
   MonitorManagementClient as GeneratedMetricsNamespacesClient
 } from "./generated/metricsnamespaces/src";
 import {
@@ -85,53 +86,193 @@ export class MetricsQueryClient {
   /**
    * Query metrics, given a resource URI
    * @param resourceUri - The resource URI to query.
-   * @param timespan - The enclosing timespan for metrics.
+   * @param metricNames - The names of the metrics to retrieve.
    * @param options - Options for querying metrics.
    * @returns A response containing metrics.
    */
-  async queryMetrics(
+  async query(
     resourceUri: string,
-    timespan: string,
-    options?: QueryMetricsOptions
-  ): Promise<QueryMetricsResult> {
+    metricNames: string[],
+    options?: MetricsQueryOptions
+  ): Promise<MetricsQueryResult> {
     const response = await this._metricsClient.metrics.list(
       resourceUri,
-      convertRequestForMetrics(timespan, options)
+      convertRequestForMetrics(metricNames, options)
     );
 
     return convertResponseForMetrics(response);
   }
 
   /**
+   * List alert segments for Metric Definitions
+   */
+  private async *listSegmentOfMetricDefinitions(
+    resourceUri: string,
+    options: ListMetricDefinitionsOptions = {}
+  ): AsyncIterableIterator<Array<MetricDefinition>> {
+    const segmentResponse = await this._definitionsClient.metricDefinitions.list(
+      resourceUri,
+      convertRequestOptionsForMetricsDefinitions(options)
+    );
+    yield convertResponseForMetricsDefinitions(segmentResponse.value);
+  }
+
+  /**
+   * List items for Metric Definitions
+   */
+  private async *listItemsOfMetricDefinitions(
+    resourceUri: string,
+    options?: ListMetricDefinitionsOptions
+  ): AsyncIterableIterator<MetricDefinition> {
+    for await (const segment of this.listSegmentOfMetricDefinitions(resourceUri, options)) {
+      if (segment) {
+        yield* segment;
+      }
+    }
+  }
+
+  /**
+   *   /**
+   *
+   * Returns an async iterable iterator to list metric definitions.
+   *
+   * Example using `for await` syntax:
+   *
+   * ```js
+   * const metricsQueryClient = new MetricsQueryClient(tokenCredential);
+   * const metricDefinitions = client.listMetricDefinitions(resourceUri, options);
+   * let i = 1;
+   * for await (const metricDefinition of metricDefinitions) {
+   *   console.log(`metricDefinition ${i++}:`);
+   *   console.log(metricDefinition);
+   * }
+   * ```
+   *
+   * Example using `iter.next()`:
+   *
+   * ```js
+   * let iter = client.listMetricDefinitions(resourceUri, options);
+   * let result = await iter.next();
+   * while (!result.done) {
+   *   console.log(` metricDefinitions - ${result.value.id}, ${result.value.name}`);
+   *   result = await iter.next();
+   * }
+   * ```
+   *
    * Get a list of metric definitions, given a resource URI.
    * @param resourceUri - The resource URI to get metric definitions for.
    * @param options - Options for getting metric definitions.
    * @returns Metric definitions for a given resource URI.
    */
-  async getMetricDefinitions(
+  listMetricDefinitions(
     resourceUri: string,
-    options?: GetMetricDefinitionsOptions
-  ): Promise<GetMetricDefinitionsResult> {
-    const response = await this._definitionsClient.metricDefinitions.list(
-      resourceUri,
-      convertRequestOptionsForMetricsDefinitions(options)
-    );
-
-    return convertResponseForMetricsDefinitions(response);
+    options?: ListMetricDefinitionsOptions
+  ): PagedAsyncIterableIterator<MetricDefinition> {
+    const iter = this.listItemsOfMetricDefinitions(resourceUri, options);
+    return {
+      /**
+       * The next method, part of the iteration protocol
+       */
+      next() {
+        return iter.next();
+      },
+      /**
+       * The connection to the async iterator, part of the iteration protocol
+       */
+      [Symbol.asyncIterator]() {
+        return this;
+      },
+      /**
+       * @returns an AsyncIterableIterator that works a page at a time
+       */
+      byPage: () => {
+        return this.listSegmentOfMetricDefinitions(resourceUri, options);
+      }
+    };
   }
 
   /**
+   * List alert segments for Metric Namespaces
+   */
+  private async *listSegmentOfMetricNamespaces(
+    resourceUri: string,
+    options: ListMetricNamespacesOptions = {}
+  ): AsyncIterableIterator<Array<MetricNamespace>> {
+    const segmentResponse = await this._namespacesClient.metricNamespaces.list(
+      resourceUri,
+      options
+    );
+    yield convertResponseForMetricNamespaces(segmentResponse.value);
+  }
+  /**
+   * List items for Metric Namespaces
+   */
+  private async *listItemsOfMetricNamespaces(
+    resourceUri: string,
+    options?: ListMetricNamespacesOptions
+  ): AsyncIterableIterator<MetricNamespace> {
+    for await (const segment of this.listSegmentOfMetricNamespaces(resourceUri, options)) {
+      if (segment) {
+        yield* segment;
+      }
+    }
+  }
+  /**
+   *
+   * Returns an async iterable iterator to list metric namespaces.
+   *
+   * Example using `for await` syntax:
+   *
+   * ```js
+   * const metricsQueryClient = new MetricsQueryClient(tokenCredential);
+   * const metricNamespaces = client.listMetricNamespaces(resourceUri, options);
+   * let i = 1;
+   * for await (const metricNamespace of metricNamespaces) {
+   *   console.log(`metricNamespace ${i++}:`);
+   *   console.log(metricNamespace);
+   * }
+   * ```
+   *
+   * Example using `iter.next()`:
+   *
+   * ```js
+   * let iter = client.listMetricNamespaces(resourceUri, options);
+   * let result = await iter.next();
+   * while (!result.done) {
+   *   console.log(` metricNamespace - ${result.value.id}, ${result.value.name}`);
+   *   result = await iter.next();
+   * }
+   * ```
    * Get a list of metric namespaces, given a resource URI.
    * @param resourceUri - The resource URI to get metric namespaces for.
    * @param options - Options for getting metric namespaces.
    * @returns Metric namespaces for a given resource URI.
    */
-  async getMetricNamespaces(
+  listMetricNamespaces(
     resourceUri: string,
-    options?: GetMetricNamespacesOptions
-  ): Promise<GetMetricNamespacesResult> {
-    const response = await this._namespacesClient.metricNamespaces.list(resourceUri, options);
-    return convertResponseForMetricNamespaces(response);
+    options?: ListMetricNamespacesOptions
+  ): PagedAsyncIterableIterator<MetricNamespace> {
+    const iter = this.listItemsOfMetricNamespaces(resourceUri, options);
+    return {
+      /**
+       * The next method, part of the iteration protocol
+       */
+      next() {
+        return iter.next();
+      },
+      /**
+       * The connection to the async iterator, part of the iteration protocol
+       */
+      [Symbol.asyncIterator]() {
+        return this;
+      },
+      /**
+       * @returns an AsyncIterableIterator that works a page at a time
+       */
+      byPage: () => {
+        return this.listSegmentOfMetricNamespaces(resourceUri, options);
+      }
+    };
   }
 }