+++ sdk/schemaregistry/schema-registry/src/models.ts
@@ -6,7 +6,7 @@ import { CommonClientOptions, OperationOptions } from "@azure/core-client";
 /**
  * Identifies a Schema by its unique ID, version, and location.
  */
-export interface SchemaId {
+export interface SchemaProperties {
   /** ID that uniquely identifies a schema in the registry namespace. */
   id: string;
 
@@ -18,12 +18,6 @@ export interface SchemaId {
 
   /** Automatically incremented version number of the schema. */
   version: number;
-
-  /** URL of schema by group and name. */
-  location: string;
-
-  /** URL of schema by ID. */
-  locationById: string;
 }
 
 /**
@@ -31,7 +25,7 @@ export interface SchemaId {
  */
 export interface SchemaDescription {
   /** Schema group under which schema is or should be registered. */
-  group: string;
+  groupName: string;
 
   /** Name of schema.*/
   name: string;
@@ -49,7 +43,7 @@ export interface SchemaDescription {
 /**
  * Schema definition with its unique ID, version, and location.
  */
-export interface Schema extends SchemaId {
+export interface Schema extends SchemaProperties {
   /** String representation of schema. */
   content: string;
 }
@@ -65,14 +59,14 @@ export interface SchemaRegistryClientOptions extends CommonClientOptions {}
 export interface RegisterSchemaOptions extends OperationOptions {}
 
 /**
- * Options for SchemaRegistryClient.getSchemaId.
+ * Options for SchemaRegistryClient.getSchemaProperties.
  */
-export interface GetSchemaIdOptions extends OperationOptions {}
+export interface GetSchemaPropertiesOptions extends OperationOptions {}
 
 /**
- * Options to configure SchemaRegistryClient.getSchemaById.
+ * Options to configure SchemaRegistryClient.getSchema.
  */
-export interface GetSchemaByIdOptions extends OperationOptions {}
+export interface GetSchemaOptions extends OperationOptions {}
 
 /**
  * Represents a store of registered schemas.
@@ -91,7 +85,10 @@ export interface SchemaRegistry {
    * @param schema - Schema to register.
    * @returns Registered schema's ID.
    */
-  registerSchema(schema: SchemaDescription, options?: RegisterSchemaOptions): Promise<SchemaId>;
+  registerSchema(
+    schema: SchemaDescription,
+    options?: RegisterSchemaOptions
+  ): Promise<SchemaProperties>;
 
   /**
    * Gets the ID of an existing schema with matching name, group, type, and
@@ -100,10 +97,10 @@ export interface SchemaRegistry {
    * @param schema - Schema to match.
    * @returns Matched schema's ID or undefined if no matching schema was found.
    */
-  getSchemaId(
+  getSchemaProperties(
     schema: SchemaDescription,
-    options?: GetSchemaIdOptions
-  ): Promise<SchemaId | undefined>;
+    options?: GetSchemaPropertiesOptions
+  ): Promise<SchemaProperties | undefined>;
 
   /**
    * Gets an existing schema by ID.
@@ -111,5 +108,5 @@ export interface SchemaRegistry {
    * @param id - Unique schema ID.
    * @returns Schema with given ID or undefined if no schema was found with the given ID.
    */
-  getSchemaById(id: string, options?: GetSchemaByIdOptions): Promise<Schema | undefined>;
+  getSchema(id: string, options?: GetSchemaOptions): Promise<Schema | undefined>;
 }